% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{combine_files_from_folder}
\alias{combine_files_from_folder}
\title{Combine CSV and TXT Files from a Folder}
\usage{
combine_files_from_folder(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing the CSV or TSV files.}
}
\value{
A data.table containing the combined data from all files.
}
\description{
This function reads all CSV and TXT files from a specified folder and combines them
into a single data.table.
}
\examples{
folder_path <- ""
combined_df <- combine_files_from_folder(folder_path)
print(combined_df)

}
