% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{run_pepmap_app}
\alias{run_pepmap_app}
\title{Launch PepMapViz Shiny Application}
\usage{
run_pepmap_app(...)
}
\arguments{
\item{...}{Additional arguments to pass to shiny::runApp()}
}
\value{
The Shiny application object
}
\description{
This function launches a Shiny application that provides an interactive
interface for the PepMapViz package functionality.
}
\examples{
\dontrun{
run_pepmap_app()
}
}
