% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_sequence.R
\name{strip_sequence_Maxquant}
\alias{strip_sequence_Maxquant}
\title{Strip sequence from Maxquant outputs}
\usage{
strip_sequence_Maxquant(data, column, convert_column)
}
\arguments{
\item{data}{A dataframe with a column containing peptide sequences to be stripped}

\item{column}{The name of the column containing the peptide sequences to be stripped.}

\item{convert_column}{The name of the column where the stripped sequences will be stored.}
}
\value{
A dataframe with a column containing stripped sequence
}
\description{
This function takes Maxquant output containing a column with peptide sequences to be stripped
and converts it into a new dataframe with the stripped sequence
}
\examples{
library(data.table)
data <- data.table(
  'Modified sequence' = c(
    "_(ac)AA(ox)AAELRLLEK_",
    "_EAAENSLVAYK_",
    "_AADTIGYPVM(ox)IRSAYALGGLGSGICPNK_"
  ),
  Condition = c("A", "B", "B")
)
column <- 'Modified sequence'
convert_column <- 'Sequence'
converted_data <- strip_sequence_Maxquant(data, column, convert_column)

}
